worker/fbposter/requirements.txt
facebook-sdk==3.1.0
python-dotenv==0.17.1